+++ src/SlackConnector.Tests.Unit/SlackConnector.Tests.Unit.csproj
@@ -80,8 +80,10 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Resources\ResourceManager.cs" />
     <Compile Include="SlackConnectorTests\Connect\ConnectedStatusTests.cs" />
+    <Compile Include="SlackConnectorTests\Connect\InboundMessageTests.cs" />
     <Compile Include="SlackConnectorTests\Connect\MultipleConnections.cs" />
     <Compile Include="SlackConnectorTests\Connect\HubsTests.cs" />
+    <Compile Include="SlackConnectorTests\Connect\Setups\ValidSetup.cs" />
     <Compile Include="SlackConnectorTests\Connect\UsersTests.cs" />
     <Compile Include="SlackConnectorTests\Connect\TeamAndSelfTests.cs" />
     <Compile Include="SlackConnectorTests\Connect\WebSocketTests.cs" />
