+++ src/SlackConnector.Tests.Unit/SlackConnectorTests/Connect/ConnectedStatusTests.cs
@@ -5,36 +5,14 @@ using Should;
 using SlackConnector.Connections;
 using SlackConnector.Connections.Handshaking;
 using SlackConnector.Connections.Handshaking.Models;
-using SlackConnector.Connections.Sockets;
-using SlackConnector.Exceptions;
-using SpecsFor;
+using SlackConnector.Tests.Unit.SlackConnectorTests.Connect.Setups;
 
 namespace SlackConnector.Tests.Unit.SlackConnectorTests.Connect
 {
     public static class ConnectedStatusTests
     {
-        public class given_valid_setup_when_connected : SpecsFor<SlackConnector>
+        public class given_valid_setup_when_connected : ValidSetup
         {
-            protected override void InitializeClassUnderTest()
-            {
-                SUT = new SlackConnector(GetMockFor<IConnectionFactory>().Object);
-            }
-
-            protected override void Given()
-            {
-                GetMockFor<IConnectionFactory>()
-                    .Setup(x => x.CreateHandshakeClient())
-                    .Returns(GetMockFor<IHandshakeClient>().Object);
-
-                GetMockFor<IHandshakeClient>()
-                    .Setup(x => x.FirmShake(It.IsAny<string>()))
-                    .ReturnsAsync(new SlackHandshake());
-
-                GetMockFor<IConnectionFactory>()
-                    .Setup(x => x.CreateWebSocketClient(It.IsAny<string>()))
-                    .Returns(GetMockFor<IWebSocketClient>().Object);
-            }
-
             protected override void When()
             {
                 SUT.Connect("key").Wait();
@@ -66,7 +44,7 @@ namespace SlackConnector.Tests.Unit.SlackConnectorTests.Connect
             }
         }
 
-        public class given_empty_api_key : SpecsFor<SlackConnector>
+        public class given_empty_api_key : ValidSetup
         {
             protected override void InitializeClassUnderTest()
             {
