+++ src/SlackConnector.Tests.Unit/SlackConnectorTests/Connect/InboundMessageTests.cs
@@ -1,0 +1,12 @@
+﻿using SpecsFor;
+
+namespace SlackConnector.Tests.Unit.SlackConnectorTests.Connect
+{
+    public static class InboundMessageTests
+    {
+        public class given_connector_is_setup_when_inbound_message_arrives : SpecsFor<SlackConnector>
+        {
+
+        }
+    }
+}
