+++ src/SlackConnector.Tests.Unit/SlackConnectorTests/Connect/MultipleConnections.cs
@@ -1,38 +1,17 @@
 ﻿using System;
-using Moq;
 using NUnit.Framework;
-using SlackConnector.Connections;
-using SlackConnector.Connections.Handshaking;
-using SlackConnector.Connections.Handshaking.Models;
-using SlackConnector.Connections.Sockets;
 using SlackConnector.Exceptions;
-using SpecsFor;
+using SlackConnector.Tests.Unit.SlackConnectorTests.Connect.Setups;
 
 namespace SlackConnector.Tests.Unit.SlackConnectorTests.Connect
 {
     public static class MultipleConnections
     {
-        public class given_connector_is_already_connected_when_calling_connect : SpecsFor<SlackConnector>
+        public class given_connector_is_already_connected_when_calling_connect : ValidSetup
         {
-            protected override void InitializeClassUnderTest()
-            {
-                SUT = new SlackConnector(GetMockFor<IConnectionFactory>().Object);
-            }
-
             protected override void Given()
             {
-                GetMockFor<IConnectionFactory>()
-                    .Setup(x => x.CreateHandshakeClient())
-                    .Returns(GetMockFor<IHandshakeClient>().Object);
-
-                GetMockFor<IHandshakeClient>()
-                    .Setup(x => x.FirmShake(It.IsAny<string>()))
-                    .ReturnsAsync(new SlackHandshake());
-
-                GetMockFor<IConnectionFactory>()
-                    .Setup(x => x.CreateWebSocketClient(It.IsAny<string>()))
-                    .Returns(GetMockFor<IWebSocketClient>().Object);
-
+                base.Given();
                 SUT.Connect("something").Wait();
             }
